+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/base/Table.java
@@ -10,6 +10,8 @@ import de.charite.compbio.jannovar.Immutable;
 @Immutable
 public final class Table {
 	private final String name;
+	private final String dbName;
+	private final String dbVersion;
 	private final String defaultPrefix;
 	private final ImmutableList<TableField> fields;
 
@@ -17,11 +19,15 @@ public final class Table {
 	 * Constructor.
 	 *
 	 * @param name Name of the table.
+	 * @param dbName Name of the database.
+	 * @param dbVersion Version of the databse.
 	 * @param defaultPrefix The default prefix to use for annotation.
 	 * @param fields The fields available in the table.
 	 */
-	public Table(String name, String defaultPrefix, Iterable<TableField> fields) {
+	public Table(String name, String dbName, String dbVersion, String defaultPrefix, Iterable<TableField> fields) {
 		this.name = name;
+		this.dbName = dbName;
+		this.dbVersion = dbVersion;
 		this.defaultPrefix = defaultPrefix;
 		this.fields = ImmutableList.copyOf(fields);
 	}
@@ -29,6 +35,10 @@ public final class Table {
 	public String getName() {
 		return name;
 	}
+
+	public String getDbName() { return dbName; }
+
+	public String getDbVersion() { return dbVersion; }
 
 	public String getDefaultPrefix() {
 		return defaultPrefix;
@@ -42,6 +52,8 @@ public final class Table {
 	public String toString() {
 		return "Table{" +
 			"name='" + name + '\'' +
+			", dbName='" + dbName + '\'' +
+			", dbVersion='" + dbVersion + '\'' +
 			", defaultPrefix='" + defaultPrefix + '\'' +
 			", fields=" + fields +
 			'}';
@@ -52,11 +64,11 @@ public final class Table {
 		if (this == o) return true;
 		if (o == null || getClass() != o.getClass()) return false;
 		Table table = (Table) o;
-		return Objects.equal(getName(), table.getName()) && Objects.equal(getDefaultPrefix(), table.getDefaultPrefix()) && Objects.equal(getFields(), table.getFields());
+		return Objects.equal(getName(), table.getName()) && Objects.equal(getDbName(), table.getDbName()) && Objects.equal(getDbVersion(), table.getDbVersion()) && Objects.equal(getDefaultPrefix(), table.getDefaultPrefix()) && Objects.equal(getFields(), table.getFields());
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hashCode(getName(), getDefaultPrefix(), getFields());
+		return Objects.hashCode(getName(), getDbName(), getDbVersion(), getDefaultPrefix(), getFields());
 	}
 }